+++ packages/manager/src/managers/telemetry/types.ts
@@ -20,6 +20,7 @@ export const SegmentEventType = {
 	customType_saved: "custom-type:saved",
 	slice_created: "slice:created",
 	identifyUser: "identify-user",
+	groupLibraries: "group-libraries",
 } as const;
 type SegmentEventTypes = typeof SegmentEventType[keyof typeof SegmentEventType];
 
@@ -52,6 +53,7 @@ export const HumanSegmentEventType = {
 	[SegmentEventType.customType_saved]: "SliceMachine Custom Type Saved",
 	[SegmentEventType.slice_created]: "SliceMachine Slice Created",
 	[SegmentEventType.identifyUser]: "IdentifyUser",
+	[SegmentEventType.groupLibraries]: "GroupLibraries",
 } as const;
 export type HumanSegmentEventTypes =
 	typeof HumanSegmentEventType[keyof typeof HumanSegmentEventType];
@@ -180,6 +182,18 @@ type IdentifyUserSegmentEvent = SegmentEvent<
 	typeof SegmentEventType.identifyUser
 >;
 
+type GroupLibrariesSegmentEvent = SegmentEvent<
+	typeof SegmentEventType.groupLibraries,
+	{
+		repoName: string;
+		manualLibsCount: number;
+		downloadedLibsCount: number;
+		npmLibsCount: number;
+		downloadedLibs: Array<string>;
+		slicemachineVersion: string;
+	}
+>;
+
 export type SegmentEvents =
 	| CommandInitStartSegmentEvent
 	| CommandInitIdentifySegmentEvent
@@ -201,4 +215,5 @@ export type SegmentEvents =
 	| CustomTypeSliceZoneUpdatedSegmentEvent
 	| CustomTypeSavedSegmentEvent
 	| SliceCreatedSegmentEvent
-	| IdentifyUserSegmentEvent;
+	| IdentifyUserSegmentEvent
+	| GroupLibrariesSegmentEvent;
