+++ packages/slice-machine/lib/models/tracking.ts
@@ -2,8 +2,6 @@
 // import { InvalidCustomTypeResponse } from "./common/TransactionalPush";
 
 export enum EventNames {
-  GroupLibraries = "GroupLibraries",
-
   ScreenshotTaken = "SliceMachine Screenshot Taken",
   ChangesPushed = "SliceMachine Changes Pushed",
   ChangesLimitReach = "SliceMachine Changes Limit Reach",
@@ -15,18 +13,6 @@ type BaseTrackingEvent = {
   name: EventNames;
   props?: Record<string, unknown>;
 };
-
-export interface GroupLibraries extends BaseTrackingEvent {
-  name: EventNames.GroupLibraries;
-  props: {
-    repoName: string;
-    manualLibsCount: number;
-    downloadedLibsCount: number;
-    npmLibsCount: number;
-    downloadedLibs: Array<string>;
-    slicemachineVersion: string;
-  };
-}
 
 export interface ScreenshotTaken extends BaseTrackingEvent {
   name: EventNames.ScreenshotTaken;
@@ -68,7 +54,6 @@ export interface EditorWidgetUsed extends BaseTrackingEvent {
 }
 
 export type TrackingEvents =
-  | GroupLibraries
   | ScreenshotTaken
   | ChangesPushed
   | ChangesLimitReach
@@ -79,10 +64,4 @@ export function isTrackingEvent(
 ): payload is TrackingEvents {
   const names = Object.values(EventNames);
   return payload.name && names.includes(payload.name);
-}
-
-export function isGroupLibrariesEvent(
-  payload: TrackingEvents
-): payload is GroupLibraries {
-  return payload.name === EventNames.GroupLibraries;
 }
