+++ packages/slice-machine/src/hooks/useSMTracker.ts
@@ -1,6 +1,5 @@
 import { useEffect } from "react";
 
-import Tracker from "@src/tracking/client";
 import { track } from "@src/apiClient";
 import { useSelector } from "react-redux";
 import { SliceMachineStoreType } from "@src/redux/type";
@@ -12,6 +11,7 @@ import {
   getIntercomHash,
 } from "@src/modules/environment";
 import { getLibraries } from "@src/modules/slices";
+import type { LibraryUI } from "@lib/models/common/LibraryUI";
 import { useRouter } from "next/router";
 
 const useSMTracker = () => {
@@ -34,7 +34,7 @@ const useSMTracker = () => {
   const router = useRouter();
 
   useEffect(() => {
-    void Tracker.get().groupLibraries(libraries, repoName, currentVersion);
+    void trackGroupLibraries(libraries, repoName, currentVersion);
 
     // For initial loading
     void trackPageView(framework, currentVersion);
@@ -65,6 +65,26 @@ const useSMTracker = () => {
 };
 
 export default useSMTracker;
+
+async function trackGroupLibraries(
+  libs: readonly LibraryUI[],
+  repoName: string | undefined,
+  version: string
+): ReturnType<typeof track> {
+  if (repoName === undefined) return;
+  const downloadedLibs = libs.filter((l) => l.meta.isDownloaded);
+  return track({
+    event: "group-libraries",
+    repoName: repoName,
+    manualLibsCount: libs.filter((l) => l.meta.isManual).length,
+    downloadedLibsCount: downloadedLibs.length,
+    npmLibsCount: libs.filter((l) => l.meta.isNodeModule).length,
+    downloadedLibs: downloadedLibs.map((l) =>
+      l.meta.name != null ? l.meta.name : "Unknown"
+    ),
+    slicemachineVersion: version,
+  });
+}
 
 function trackPageView(
   framework: string,
